Adds parallel cyclic reduction (PCR) as an option for solving batched time systems (#25)
Parallel cyclic reduction in theory has quite good scaling/vectorization on the GPU -- better than Thomas's algorithm anyway. In practice it seems to perform well for problems with a small number of coupled ODEs. This PR adds the algorithm and updates some of the code infrastructure to switch between different linear solver approaches.